publish.sources: fixed rsync to prevent runaway upload
encyc.rsync.push was intended to rsync a list of files.
encyc.publish.sources was calling it once per file, passing a str.
This caused rsync to sync a bunch of files in / which was *not*
what was intended.
This fix constructs the list and feeds it to encyc.rsync.push.
Also, replaced hard-coded URL for PSMS export csv with config.