mark_offsite_encyc_rg_links: no proto/domains for encycfront links
Previously the function rewrote links to the main encyclopedia
with the encycfront domain, adding the protocol ('http://') and
domain.  When the function was run subsequently, these links
were understood by the function to be offsite links, the markers
were removed, and subsequent passes would fail to mark links
properly.

This change marks links as rg/notrg but does not modify the links
themselves.  It is the consuming app's (e.g. encycrg) job to
rewrite the links.